Added 'default' values to the metamodel.
I have added a new annotation, @default(VALUE), in order to assign a
default value for an attribute. Since we cannot annotate the value
itself (because it is attached to the real field, never an accessor)
the annotation must go with the type (AnnotatedType):

trait A {
  val a1: String @default("")
  val a2: Int @default(5)
}

If we try to access the default value ('default' method at MetaModel)
and it is found a value that does not conform the attribute type, an
expection will be arised.

There is still a lot of work to go, because currently the MetaModel
can only understand constant literals.