Default values at annotations are now parsed using 'eval'.
The runtime metamodel is now able to evaluate the annotation argument
(remember: it's a Tree!). In order to do so, I have added a tool-box
in the runtime metamodel, which contains the eval method. I think that
the implementation may be moved to the generic metamodel, but it
involves adding new concepts such as mirror and toolbox.