Now builder[A] knows how to handle @default(value).
We had added @default(values) in previous commits, and know I have
tought the builder about this issue. If an attribute has a @default
then it is used as default value. However, if the annotation is not
found, then the Default type class does its job.

By introducing this concept in the builder, we can detect invalid
default types in compilation time. This rocks!