FIXED: not every builder belongs to a refined type.
I have seen that there are some situations where a builder in the
system doesn't belong to a refined type. Instead, you can find a
member in the system whose type is just Builder[T]. In those cases the
finder was failing. A patch has been added to solve the issue.

Besides, I have found 'Null' builders (dummyX), which made the finder
fail, and now they are ignored.