Added an "artificial" extractor for attributes.
While working with Speech we saw that dealing with _Set events was a
little bit annoying, so we thought an extractor at attributes would be
a nice idea. However, we didn't want to add speech dependencies
inside updatable. To do so, we have used an artificial 'Extractee', a
structural type that contains the basic methods to extract the info.