Added a method to extract attributes whose type is collection.
Due to new requirements in the Speech Javascript Generation, I have
added a new method in the metamodel in order to extract all the
attributes whose type is Collection (Set or List).

I have found that I can't invoke typeOf inside the macro metamodel so
I had to compare the types against the string, and that is kind of
ugly.